requirements.txt
ollama
torch
python-dotenv
Pillow
transformers
